# agent/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "随心说 Language Processing API",
    "version": "2.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/polish": {
      "post": {
        "summary": "Polish Endpoint",
        "operationId": "polish_endpoint_v1_polish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-remaining-budget-ms",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "description": "可选：客户端发送时剩余预算；手机仍须执行从 FINAL 起算的总期限。",
              "default": 10000,
              "title": "X-Remaining-Budget-Ms"
            },
            "description": "可选：客户端发送时剩余预算；手机仍须执行从 FINAL 起算的总期限。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolishResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    },
    "/v1/compose-signs": {
      "post": {
        "summary": "Compose Signs Endpoint",
        "operationId": "compose_signs_endpoint_v1_compose_signs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-remaining-budget-ms",
            "in": "header",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 1,
              "description": "客户端剩余处理预算，单位毫秒；云端调用最多 10 秒。",
              "default": 10000,
              "title": "X-Remaining-Budget-Ms"
            },
            "description": "客户端剩余处理预算，单位毫秒；云端调用最多 10 秒。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignComposeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignComposeResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ContextSentence": {
        "properties": {
          "segmentId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Segmentid"
          },
          "rawChinese": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Rawchinese"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segmentId",
          "rawChinese"
        ],
        "title": "ContextSentence"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "segmentId": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmentid"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2147483647.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "FidelityIssue": {
        "properties": {
          "language": {
            "type": "string",
            "pattern": "^[a-z]{2,3}(?:-[A-Z][a-z]{3})?(?:-[A-Z]{2}|-[0-9]{3})?$",
            "title": "Language"
          },
          "code": {
            "type": "string",
            "enum": [
              "AMBIGUITY",
              "FIDELITY_CHECK_FAILED",
              "UNAVAILABLE"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "language",
          "code",
          "message"
        ],
        "title": "FidelityIssue"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status",
            "default": "ok"
          },
          "model_configured": {
            "type": "boolean",
            "title": "Model Configured"
          }
        },
        "type": "object",
        "required": [
          "model_configured"
        ],
        "title": "HealthResponse"
      },
      "PolishRequest": {
        "properties": {
          "segmentId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Segmentid"
          },
          "rawChinese": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Rawchinese"
          },
          "sessionId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Sessionid"
          },
          "revision": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Revision"
          },
          "targetLanguages": {
            "items": {
              "type": "string",
              "pattern": "^[a-z]{2,3}(?:-[A-Z][a-z]{3})?(?:-[A-Z]{2}|-[0-9]{3})?$"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Targetlanguages"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/ContextSentence"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segmentId",
          "rawChinese",
          "sessionId",
          "revision"
        ],
        "title": "PolishRequest"
      },
      "PolishResponse": {
        "properties": {
          "polishedChinese": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Polishedchinese"
          },
          "translations": {
            "patternProperties": {
              "^[a-z]{2,3}(?:-[A-Z][a-z]{3})?(?:-[A-Z]{2}|-[0-9]{3})?$": {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              }
            },
            "type": "object",
            "title": "Translations"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/FidelityIssue"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Issues"
          },
          "segmentId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Segmentid"
          },
          "revision": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "polishedChinese",
          "translations",
          "issues",
          "segmentId",
          "revision"
        ],
        "title": "PolishResponse"
      },
      "SignCandidate": {
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "一定",
              "你",
              "只",
              "可以",
              "回",
              "大家",
              "好",
              "家",
              "年",
              "很久不",
              "我",
              "我们",
              "新",
              "是",
              "照顾",
              "祝贺",
              "自己",
              "要",
              "见"
            ],
            "title": "Label"
          },
          "score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score",
            "description": "未校准的 CV softmax 分数，不是可靠概率"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label"
        ],
        "title": "SignCandidate"
      },
      "SignComposeRequest": {
        "properties": {
          "sessionId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Sessionid"
          },
          "segmentId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Segmentid"
          },
          "revision": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Revision"
          },
          "gestures": {
            "items": {
              "$ref": "#/components/schemas/SignGesture"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 1,
            "title": "Gestures"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessionId",
          "segmentId",
          "revision",
          "gestures"
        ],
        "title": "SignComposeRequest"
      },
      "SignComposeResponse": {
        "properties": {
          "segmentId": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Segmentid"
          },
          "revision": {
            "type": "integer",
            "maximum": 2147483647.0,
            "minimum": 0.0,
            "title": "Revision"
          },
          "sentence": {
            "type": "string",
            "enum": [
              "我想回家",
              "你一定可以",
              "你要照顾好自己",
              "祝大家新年好",
              "我们只是好久不见"
            ],
            "title": "Sentence"
          },
          "alternatives": {
            "items": {
              "type": "string",
              "enum": [
                "我想回家",
                "你一定可以",
                "你要照顾好自己",
                "祝大家新年好",
                "我们只是好久不见"
              ]
            },
            "type": "array",
            "title": "Alternatives"
          },
          "status": {
            "type": "string",
            "const": "CANDIDATE",
            "title": "Status"
          },
          "needsConfirmation": {
            "type": "boolean",
            "const": true,
            "title": "Needsconfirmation",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segmentId",
          "revision",
          "sentence",
          "alternatives",
          "status"
        ],
        "title": "SignComposeResponse"
      },
      "SignGesture": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/SignCandidate"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Candidates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidates"
        ],
        "title": "SignGesture"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}